+++ docs/src/assets/logo.png
