+++ docs/src/index.html
@@ -7,10 +7,10 @@
   <meta name="author" content="Allan Alexandre">
   <meta name="description" content="vishnucss provides a minimal setup of styles for a fast and clean starting point.">
   <title>Vishnu - A simple and minimalist toolkit with modern css.</title>
-  <link rel="icon" href="./logo.png">
-  <link rel="stylesheet" href="./vishnu.min.css">
-  <link rel="stylesheet" href="./vishnu.extensions.min.css">
-  <link rel="stylesheet" href="./style.css">
+  <link rel="icon" href="./assets/logo.png">
+  <link rel="stylesheet" href="./assets/vishnu.min.css">
+  <link rel="stylesheet" href="./assets/vishnu.extensions.min.css">
+  <link rel="stylesheet" href="./assets/style.css">
 </head>
 
 <body>
@@ -19,11 +19,11 @@
     <a href="https://github.com/vishnucss/vishnu"><img style="position: absolute; top: 0; right: 0; border: 0;" src="https://camo.githubusercontent.com/38ef81f8aca64bb9a64448d0d70f1308ef5341ab/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f6461726b626c75655f3132313632312e706e67" alt="Fork me on GitHub" data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_darkblue_121621.png"></a>
     <section class="hero flex center">
       <div class="text-center">
-        <img src="./logo.png" alt="Vishnu Logo" height="180">
+        <img src="./assets/logo.png" alt="Vishnu Logo" height="180">
         <h1>Vishnu</h1>
         <h5>A simple and minimalist toolkit with modern css</h5>
         <a class="github-button" href="https://github.com/vishnucss/vishnu" data-icon="octicon-star" data-size="large" aria-label="Star vishnucss/vishnu on GitHub">Star</a>
-        <small class="full-width">Currently v1.0.0-beta.7</small>
+        <small class="full-width">Currently v1.0.0-beta.8</small>
       </div>
     </section>
 
@@ -87,22 +87,6 @@
           </div>
           <div class="row">
             <div class="col-xs-4 col-sm-12">
-              <button class="large">Large Button</button>
-            </div>
-            <div class="col-xs-8 col-sm-12">
-              <p>This button is available by class <code>.large</code></p><code>&lt;button class='large'&gt;Content&lt;/button&gt;</code>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-xs-4 col-sm-12">
-              <button class="small">Small Button</button>
-            </div>
-            <div class="col-xs-8 col-sm-12">
-              <p>This button is available by class <code>.small</code></p><code>&lt;button class='small'&gt;Content&lt;/button&gt;</code>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col-xs-4 col-sm-12">
               <button class="outline">Outlined Button</button>
             </div>
             <div class="col-xs-8 col-sm-12">
@@ -133,6 +117,25 @@
               <p>This button is available with a single class</p><code>&lt;button class='rounded'&gt;content&lt;/button&gt;</code>
             </div>
           </div>
+
+          <h4>Sizes</h4>
+          <div class="row">
+            <div class="col-xs-4 col-sm-12">
+              <button class="large">Large Button</button>
+            </div>
+            <div class="col-xs-8 col-sm-12">
+              <p>This button is available by class <code>.large</code></p><code>&lt;button class='large'&gt;Content&lt;/button&gt;</code>
+            </div>
+          </div>
+          <div class="row">
+            <div class="col-xs-4 col-sm-12">
+              <button class="small">Small Button</button>
+            </div>
+            <div class="col-xs-8 col-sm-12">
+              <p>This button is available by class <code>.small</code></p><code>&lt;button class='small'&gt;Content&lt;/button&gt;</code>
+            </div>
+          </div>
+
         </div>
 
         <!-- List -->
