+++ lerna.json
@@ -1,0 +1,7 @@
+{
+  "lerna": "2.4.0",
+  "packages": [
+    "packages/*"
+  ],
+  "version": "0.0.0"
+}
