+++ packages/vishnu/README.md
@@ -1,0 +1,35 @@
+<p align="center"><a href="https://vishnucss.github.io/vishnu" target="_blank"><img src="https://vishnucss.github.io/vishnu/logo.png" width="59" height="59" /></a></p>
+<h1 align="center">Vishnu</h1>
+
+<p align="center">
+<a href="https://circleci.com/gh/alexandesigner/vishnu/tree/master"><img src="https://circleci.com/gh/alexandesigner/vishnu/tree/master.svg?&style=shield" alt="circle ci" height="24"></a>
+<a href="https://badge.fury.io/js/vishnucss"><img src="https://badge.fury.io/js/vishnucss.svg" alt="npm version" height="24"></a>
+</p>
+
+## Why should I use it?
+
+A simple style guide provides a setup of styles using Modern CSS for a fast and clean starting point. Everything you need to create a basic project is included.
+
+## Install
+
+**Using npm**
+
+```sh
+$ npm install --save vishnucss
+```
+
+**Using directly link**
+
+```html
+<link rel="stylesheet" href="https://unpkg.com/vishnucss"/>
+```
+
+## Bug?
+
+Submit it in the [issues](https://github.com/vishnucss/vishnu/issues)
+
+## Get Involved
+
+We'd love for you to help us build Vishnu. If you'd like to be a contributor, check out our <a href="https://github.com/vishnucss/vishnucss/blob/master/.github/CONTRIBUTING.md" target="_blank">Contributing guide</a>
+
+<p>Designed with ♥ by <a target="_blank" href="http://alexandesigner.com.br" title="Allan Alexandre">Allan Alexandre</a>. Licensed under the <a target="_blank" href="https://github.com/vishnucss/vishnu#license" title="MIT License">MIT License</a>.</p>
