+++ packages/vishnucss-extensions/.versions
@@ -1,0 +1,2 @@
+meteor@1.7.2
+vishnucss:vishnucss-extensions@1.0.0-beta.0
