+++ packages/vishnucss-extensions/README.md
@@ -1,0 +1,26 @@
+<p align="center"><a href="https://vishnucss.github.io/vishnu" target="_blank"><img src="https://vishnucss.github.io/vishnu/logo.png" width="59" height="59" /></a></p>
+<h1 align="center">Vishnu Extensions</h1>
+
+## Install
+
+**Using npm**
+
+```sh
+$ npm install --save vishnucss-extensions
+```
+
+**Using directly link**
+
+```html
+<link rel="stylesheet" href="https://unpkg.com/vishnucss-extensions"/>
+```
+
+## Bug?
+
+Submit it in the [issues](https://github.com/vishnucss/vishnu/issues)
+
+## Get Involved
+
+We'd love for you to help us build Vishnu. If you'd like to be a contributor, check out our <a href="https://github.com/vishnucss/vishnucss/blob/master/.github/CONTRIBUTING.md" target="_blank">Contributing guide</a>
+
+<p>Designed with ♥ by <a target="_blank" href="http://alexandesigner.com.br" title="Allan Alexandre">Allan Alexandre</a>. Licensed under the <a target="_blank" href="https://github.com/vishnucss/vishnu#license" title="MIT License">MIT License</a>.</p>
