+++ packages/vishnucss-extensions/src/customs.css
@@ -1,0 +1,39 @@
+/*
+* Responsive
+*/
+
+@custom-media --phone-viewport only screen and (max-width: 25em);
+@custom-media --tablet-viewport only screen and (max-width: 48em);
+@custom-media --desktop-viewport only screen and (max-width: 62em);
+@custom-media --desktop-large-viewport only screen and (max-width: 75em);
+
+/*
+* Alert
+*/
+
+@custom-selector :--alert .alert;
+@custom-selector :--placeholder .placeholder;
+@custom-selector :--info .info;
+@custom-selector :--danger .danger;
+@custom-selector :--success .success;
+@custom-selector :--rounded .rounded;
+@custom-selector :--close .close;
+
+/*
+* Card
+*/
+
+@custom-selector :--card .card;
+@custom-selector :--header .card-header;
+@custom-selector :--footer .card-footer;
+@custom-selector :--title .card-title;
+@custom-selector :--content .card-content;
+
+/* 
+ * Grid
+*/
+
+@custom-selector :--all-cols-xs .col-xs, .col-xs-1, .col-xs-2, .col-xs-3, .col-xs-4, .col-xs-5, .col-xs-6, .col-xs-7, .col-xs-8, .col-xs-9, .col-xs-10, .col-xs-11, .col-xs-12;
+@custom-selector :--all-cols-sm .col-sm, .col-sm-1, .col-sm-2, .col-sm-3, .col-sm-4, .col-sm-5, .col-sm-6, .col-sm-7, .col-sm-8, .col-sm-9, .col-sm-10, .col-sm-11, .col-sm-12;
+@custom-selector :--all-cols-md .col-md, .col-md-1, .col-md-2, .col-md-3, .col-md-4, .col-md-5, .col-md-6, .col-md-7, .col-md-8, .col-md-9, .col-md-10, .col-md-11, .col-md-12;
+@custom-selector :--all-cols-lg .col-lg, .col-lg-1, .col-lg-2, .col-lg-3, .col-lg-4, .col-lg-5, .col-lg-6, .col-lg-7, .col-lg-8, .col-lg-9, .col-lg-10, .col-lg-11, .col-lg-12;
